backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-dotenv
